Compile with Gem::Ext::ExtConfBuilder.
This keeps with the "shoe philosophy" of doing things just like
Rubygems, so that development is an accurate representative of
deployment.

This change effectively adds a "make install" step to the build process,
copying the generated .so/.bundle/.dll into the lib directory, so we
don't need to bother with including ext in the LOAD_PATH anymore.