Task load order no longer matters.
Rake prerequisites are now explicitly pushed onto the end of the list or
unshifted onto the beginning.

Code-running tasks that might depend on extension compilation depend on
an arbitrary prepare task, and compile registers itself as a
prerequisite of prepare as needed.

Now I can document using Shoe tasks a la carte.