Remove shoe generator from cucumber and test tests
Also, falling back to bundler-suggested use of git in gemspec, so as to
only include version-controlled files in the gem. I think that'll be
good long-term, though it does mean we need to git add new files before
we can compile or test them.

I'm not going to think about it too much for now, since shoe's going to
be relinquishing its opinions about what gemspecs should look like
anyway.